pyproject.toml
[tool.poetry]
name = "esm"
version = "3.2.0"
description = "EvolutionaryScale open model repository"
authors = ["EvolutionaryScale Team"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/Zahid8/ESM3-Adapters"
repository = "https://github.com/Zahid8/ESM3-Adapters"
keywords = ["protein language model", "esm", "bioinformatics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.10",
]
packages = [{include = "esm", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10"
torch = ">=2.2.0"
torchvision = "*"
torchtext = "*"
transformers = "<4.48.2"
ipython = "*"
einops = "*"
biotite = ">=1.0.0"
msgpack-numpy = "*"
biopython = "*"
scikit-learn = "*"
brotli = "*"
attrs = "*"
pandas = "*"
cloudpathlib = "*"
tenacity = "*"
zstd = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"
hydra-core = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]